+++ apps/cowswap-frontend/src/pages/AdvancedOrders/index.tsx
@@ -1,4 +1,5 @@
 import { useAtomValue } from 'jotai'
+import { ReactNode } from 'react'
 
 import { PAGE_TITLES } from '@cowprotocol/common-const'
 
@@ -30,10 +31,7 @@ import { TwapFormState } from 'modules/twap/pure/PrimaryActionButton/getTwapForm
 
 const ADVANCED_ORDERS_MAX_WIDTH = '1800px'
 
-// TODO: Break down this large function into smaller functions
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export default function AdvancedOrdersPage() {
+export default function AdvancedOrdersPage(): ReactNode {
   const { isUnlocked } = useAtomValue(advancedOrdersAtom)
   const { ordersTableOnLeft } = useAtomValue(limitOrdersSettingsAtom)